Update license
This change to the terms makes it forbidden to use zVirtualDesktop on a computer owned by an organization without first obtaining approval from me. Likely I will ask for a donation based on the number of computers. At this point the plan is to keep the personal use free for everyone. The terms have always stated that zVirtualDesktop cannot be bundled with other applications and may only be distributed in a zip file that includes the license (License.txt).